初期データ投入

+++ db/sql/V2.0__create_table.sql
@@ -1,10 +1,10 @@
 -- 販売データテーブル作成
 CREATE TABLE IF NOT EXISTS sales_data(
-sales_id SERIAL NOT NULL PRIMARY KEY,
-column_number INT NOT NULL,
-sales_juice_name VARCHAR(50) NOT NULL,
-sales_price MONEY NOT NULL,
-sales_datetime TIMESTAMP,
-FOREIGN KEY(column_number) REFERENCES vending_machine_column(column_number)
+    sales_id SERIAL NOT NULL PRIMARY KEY,
+    column_number INT NOT NULL,
+    sales_juice_name VARCHAR(50) NOT NULL,
+    sales_price INT NOT NULL,
+    sales_datetime TIMESTAMP,
+    FOREIGN KEY(column_number) REFERENCES vending_machine_column(column_number)
 );
